+++ packages/r/openproblems.utils/R/validate_object.R
@@ -31,18 +31,19 @@
 #' )
 #' validate_object(task_config, what = "task_config")
 validate_object <- function(
-    obj,
-    what = c(
-      "api_component_spec",
-      "api_file_format",
-      "task_config",
-      "task_control_method",
-      "task_method",
-      "task_metric"
-    ),
-    obj_source = NULL,
-    engine = c("ajv", "imjv"),
-    error = TRUE) {
+  obj,
+  what = c(
+    "api_component_spec",
+    "api_file_format",
+    "task_config",
+    "task_control_method",
+    "task_method",
+    "task_metric"
+  ),
+  obj_source = NULL,
+  engine = c("ajv", "imjv"),
+  error = TRUE
+) {
   what <- match.arg(what)
   engine <- match.arg(engine)